게시물 수정, 삭제시 권한 체크

+++ src/main/java/com/sbs/jdbc/text_board/boundedContext/article/controller/ArticleController.java
@@ -87,6 +87,11 @@ public class ArticleController {
   }
 
   public void doModify(Rq rq) {
+    if(rq.isLogouted()) {
+      System.out.println("로그인 후 이용해주세요.");
+      return;
+    }
+
     int id = rq.getIntParam("id", 0);
 
     if(id == 0) {
@@ -95,9 +100,16 @@ public class ArticleController {
     }
 
     Article article = articleService.findById(id);
+    Member member = rq.getLoginedMember(); // 로그인한 사용자의 대한 정보를 세션에서 가져옴
 
     if(article == null) {
       System.out.printf("%d번 게시물은 존재하지 않습니다.\n", id);
+      return;
+    }
+
+    if(article.getMemberId() != member.getId()) {
+      System.out.println("해당 게시물에 대한 접근 권한이 없습니다.");
+      return;
     }
 
     System.out.printf("== %d번 게시물 수정 ==\n", id);
@@ -113,6 +125,11 @@ public class ArticleController {
   }
 
   public void doDelete(Rq rq) {
+    if(rq.isLogouted()) {
+      System.out.println("로그인 후 이용해주세요.");
+      return;
+    }
+
     int id = rq.getIntParam("id", 0);
 
     if (id == 0) {
@@ -121,9 +138,16 @@ public class ArticleController {
     }
 
     Article article = articleService.findById(id);
+    Member member = rq.getLoginedMember();
 
     if(article == null) {
       System.out.printf("%d번 게시물은 존재하지 않습니다.\n", id);
+      return;
+    }
+
+    if(article.getMemberId() != member.getId()) {
+      System.out.println("해당 게시물에 대한 접근 권한이 없습니다.");
+      return;
     }
 
     articleService.delete(id);
